Add files via upload
This script posts .csv file format documents to solr for indexing. Script can be used for load testing of solr as it can be run for specified number of seconds with specified number of threads.
Place the script in the same directory as the *.csv files & run it.

How to run:
python3 solr_post.py <num of threads> <time to run in seconds>
arg1 & arg2 are optional.